#25 moving actions inside actionCreators
