--- src/store/actions.js
@@ -1,11 +1,0 @@
-const actions = {
-	UPDATE_LOAN: 'UPDATE_LOAN',
-	UPDATE_WITH_INSURANCE: 'UPDATE_WITH_INSURANCE',
-	UPDATE_MONTHS: 'UPDATE_MONTHS',
-	CALC_LOAN_SEND: 'CALC_LOAN_SEND',
-	UPDATE_SUMMARY_FIELDS: 'UPDATE_SUMMARY_FIELDS',
-	CALC_LOAN_SEND_DEBOUNCE: 'CALC_LOAN_SEND_DEBOUNCE',
-	QUERY_CANCELLED: 'QUERY_CANCELLED',
-};
-
-export default actions;
