+++ src/store/reducers/summary-loan.js
@@ -1,4 +1,4 @@
-import actions from '../actions';
+import { actions } from '../actionCreators';
 
 const { UPDATE_SUMMARY_FIELDS } = actions;
