+++ src/store/reducers/time-months.js
@@ -1,4 +1,4 @@
-import actions from '../actions';
+import { actions } from '../actionCreators';
 
 const { UPDATE_MONTHS } = actions;
